Streamlined design + support for multiple areas.
Converted a number of fields to numbers (startTime, endTime, costPerHour, duration and disability factor) of integer and double as that is the best way to represent such data and upon serialization this data would otherwise have to be converted manually anyway.

Changed "coordinates" into an object that strongly represents longitude and latitude GPS coordinates.

Removed "numberOfProfiles" for use of an array to represent the same data. The number is therefore reflected by the number of elements in the array.

Removed the "metered" boolean as this is implicitly represented by the "duration" (i.e. if the duration is 0 or omitted then parking is unmetered during the time of the profile.

Removed "applyOnPH" as there was no description as to what this field is for.

Removed "typeDescription" as there was no description as to what this field is for. The given value "parking_restriction_data" was implicit.